Automatically restore focus on the popup source element when user tabs out of material popup.
PiperOrigin-RevId: 200091764
