+++ lib/material_select/material_dropdown_select.html
@@ -37,8 +37,7 @@
     [autoDismiss]="autoDismiss"
     [trackLayoutChanges]="trackLayoutChanges"
     [visible]="visible"
-    (visibleChange)="onVisible($event)"
-    (autoDismissed)="onAutoDismissed($event)">
+    (visibleChange)="onVisible($event)">
   <div header
       (keydown)="onKeyDown($event)"
       (keypress)="onKeyPress($event)"
