+++ lib/src/laminate/popup/dom_popup_source.dart
@@ -96,6 +96,11 @@ class DomPopupSource implements ElementPopupSource {
   }
 
   @override
+  void focus() {
+    sourceElement.focus();
+  }
+
+  @override
   String toString() =>
       'DomPopupSource ' +
       {'alignOriginX': alignOriginX, 'alignOriginY': alignOriginY}.toString();
